added missing renames
https://coveord.atlassian.net/browse/KIT-2680
