+++ packages/headless/src/app/ssr-engine/types/fetch-static-state.ts
