+++ packages/headless/src/app/ssr-engine/types/hydrate-static-state.ts
